+++ tests/capi/doc/chfl_cell/set_angles.c
@@ -7,11 +7,11 @@
 
 int main() {
     // [example]
-    CHFL_CELL* cell = chfl_cell_triclinic((chfl_vector_t){10, 10, 10}, (chfl_vector_t){90, 90, 90});
+    CHFL_CELL* cell = chfl_cell_triclinic((chfl_vector3d){10, 10, 10}, (chfl_vector3d){90, 90, 90});
 
-    chfl_cell_set_angles(cell, (chfl_vector_t){120, 110, 100});
+    chfl_cell_set_angles(cell, (chfl_vector3d){120, 110, 100});
 
-    chfl_vector_t angles = {0, 0, 0};
+    chfl_vector3d angles = {0, 0, 0};
     chfl_cell_angles(cell, angles);
     assert(angles[0] == 120);
     assert(angles[1] == 110);