+++ tests/capi/doc/chfl_frame/set_cell.c
@@ -7,7 +7,7 @@
 int main() {
     // [example]
     CHFL_FRAME* frame = chfl_frame();
-    CHFL_CELL* cell = chfl_cell((chfl_vector_t){10, 10, 12});
+    CHFL_CELL* cell = chfl_cell((chfl_vector3d){10, 10, 12});
 
     chfl_frame_set_cell(frame, cell);