+++ tests/capi/doc/chfl_trajectory/set_cell.c
@@ -6,7 +6,7 @@
 int main() {
     // [example] [no-run]
     CHFL_TRAJECTORY* trajectory = chfl_trajectory_open("water.xyz", 'r');
-    CHFL_CELL* cell = chfl_cell((chfl_vector_t){22, 22, 34});
+    CHFL_CELL* cell = chfl_cell((chfl_vector3d){22, 22, 34});
 
     chfl_trajectory_set_cell(trajectory, cell);